+++ failfast/src/test/kotlin/failfast/TestFinderTest.kt
@@ -1,5 +1,6 @@
 package failfast
 
+import failfast.docs.ClassTestContextTest
 import failfast.docs.ObjectMultipleContextsTest
 import failfast.docs.ObjectTestContextTest
 import failfast.docs.TestContextExampleTest
@@ -16,6 +17,7 @@ object TestFinderTest {
             it("can find Test classes") {
                 expectThat(FailFast.findTestClasses(classIncludeRegex = Regex(".*docs.*Test.class\$")))
                     .containsExactlyInAnyOrder(
+                        ClassTestContextTest::class,
                         ObjectTestContextTest::class,
                         TestFinderTest::class.java.classLoader.loadClass("failfast.docs.TestContextOnTopLevelTest").kotlin,
                         ObjectMultipleContextsTest::class,