+++ failfast/src/test/kotlin/failfast/docs/ClassTestContextTest.kt
@@ -1,0 +1,14 @@
+package failfast.docs
+
+import failfast.describe
+import org.junit.platform.commons.annotation.Testable
+import strikt.api.expectThat
+import strikt.assertions.isEqualTo
+class ClassTestContextTest {
+    val context =
+        describe("test context defined in a kotlin class") {
+            it("describes behavior") {
+                expectThat("test").isEqualTo("test")
+            }
+        }
+}